many general improvements - abstracted project view away to its own reusable view component - added user management tools - added invite code issuing - added new user flags for permission control - added new "account deactivated" page for bedroom - form errors now display on the nav loader - cleaned up bedroom homepage - added spoiler visuals for inputs with sensitive info - implemented sveltekit form enhancement on project and post creation modals - added link to co-authoring notice - removed unused assets
and finally
- unlocked view source button

+++ src/routes/(main)/bedroom/(colored)/posts/[id]/+page.server.ts
@@ -18,6 +18,15 @@ export const load: PageServerLoad = async (event) => {
 
 export const actions: Actions = {
   save: async (event) => {
+    const user = event.locals.user;
+    if (!user) return fail(401);
+
+    const post = (await db.select().from(table.post).where(eq(table.post.id, event.params.id))).at(0);
+    if (!post) return fail(404, { error: "PostNotFound", message: "Post not found" });
+
+    const canEdit = user.flags.includes("manage-posts:all") || (user.flags.includes("manage-posts") && post.authorId === user.id);
+    if (!canEdit) return fail(403);
+
     const formData = await event.request.formData();
 
     const title = formData.get("title") as string;
@@ -51,6 +60,15 @@ export const actions: Actions = {
     };
   },
   delete: async (event) => {
+    const user = event.locals.user;
+    if (!user) return fail(401);
+
+    const post = (await db.select().from(table.post).where(eq(table.post.id, event.params.id))).at(0);
+    if (!post) return fail(404, { error: "PostNotFound", message: "Post not found" });
+
+    const canDelete = user.flags.includes("manage-posts:all") || (user.flags.includes("manage-posts") && post.authorId === user.id);
+    if (!canDelete) return fail(403);
+
     await db.delete(table.post).where(eq(table.post.id, event.params.id));
     return redirect(303, "/bedroom/posts");
   }
